HID: multitouch: handle new multitouch sensor of Wacom tablets through hid-mt
This allows to dynamically blacklist or not the HID device and decide that
if it can be handled through hid-mt, then it will be. If the HID collection
refers to the pen or a non standard touch interface, then reject it from
the HID subsystem and use the wacom.ko module instead.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>