HID: wacom: generic HID Wacom handling
Wacom reports present several points which are slightly not accepted
by the HID protocol.
Thus, we introduce a few specific handling for them:

We first introduce a vendor specific usage page Wacom which contains
some usages not defined by the spec:
- local ang global sub report ID
- Serial number of teh tool in use
- Tool ID

Global sub report ID is here because the report may contain different
usages depending on this field for the same report ID.
For instance, report ID 2 cointains on teh Cintiq 22HDT either:
- entering in proximity report, which contains tool info (serial and id)
- actual pen data (with different usages accross the report)
- outgoing of proximity

The Global sub report ID is interpreted by raw_event() which simply
replace the current report ID before hid-core interprets its value.

The local sub report ID is not implemented right now but allows to
split the current report in several sub reports of a given size.
This allows to have a report which contains several sub reports, each
of them can have a different meaning depending on the sub report ID.

The other problem with the Wacom device is that the usages are not
packed accross the report, and some device uses big endian instead of
little endian as specified by the HID specification.
The extension proposed here is to had several vendor usage pages, each
of them prefixed by 0xbe, and the LSB is a current normalized usage page.

For instance, the Digitizer Usage Page (0x000d) becomes 0xbe0d with this
vendor collection. Once we have an usage in this vendor page, we also
split the 2 bytes of the usage in 2:
- the MSB reflects the offset of the value within the usage
- the LSB reflects the actual usage in the original page

So if we encounter a usage page of 0xbe01 with a usage 0x0930, that mean
that we have the usage page Generic Desktop with an offset, and the usage
is X which will be shifted by 9 to the left (X << 9).

This way, we can describe several parts of a usage and get the full value
just by shifting the bits accordingly.

The rest of the patch is generic accross HID drivers (add input_mapping(),
input_feature_mapping(), etc...).

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>