HID: Wacom: Add HID_WAC_INTUOS_SUB_REPORT_ID usage
The second bit of the Intuos packet tell which tool is in use, and which
type of packet will be reported.
It is not possible to simply apply a mask on the byte to get the right
information, so implementing a small function to get the actual
type of the report.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>