Input - wacom: check for valid devices based on the packet size
The Intuos Pro shares three HID device in its internal hub. Two of them
are the touch interface and the pen interface. However a third boot mouse
interface exists too. The Previous USB_DEVICE_DETAILED macro in wacom_wac.c
used to filter them, but the switch to the HID subsystem removed this
capability.

We already have a pretty big list of packet size regarding the pen
of each Wacom tablet. We can use this for checking if the current hid
device is a valid pen or not.

For touch devices, some do not have a listed pktlen, and some are sharing
the same usb PID with the pen interface. Consider the list of possible
touch packet length and validate accordingly.

Given that the pktlen is correctly computed/validated, we can store it now
in the features struct instead of having a special handling in the rest of
the code.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>