Input - wacom: remove usb dependency for siblings devices
Waom tablets can share different physical sensor on one physical device.
This are called siblings in the code. The current way of implementation
relies on the USB topology to be able to share data amongs those sensors.

We can replace the code to match a HID subsystem, without USB involved:
- the first probed sensor does not find any siblings in the list
  wacom_udev_list, so it creates its own wacom_hdev_data with its own
  struct hid_device
- the other sensor check the current list of siblings in wacom_hdev_data,
  and if he has a match, it associates itself to the matched device.

WIP: check the physical path for devices having an internal hub

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>