HID: uhid: add and set HID_TYPE_UHID for uhid devices
Some wacom devices checks on the .type field of hid_device. However,
uhid never set it, and as a result, we can not use uhid with some of
the wacom tablets.

Create a HID type HID_TYPE_UHID which is set by uhid and taken into
account by wacom.ko.

With this patch, the wacom.ko driver is completely agnostic of the
USB layer, except for the Wireless Receiver. That means that starting
from now, we can check every changes through uhis and comparing the
resulting kernel device.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>