Input - wacom: handle Intuos 4 BT in wacom.ko
Now that wacom is a hid driver, there is no point in having a separate
driver for bluetooth devices.

Another good point of this is that now, the Intuos4 bluetooth can handle
the different tools (artpen, airbrush), and we get a common interface
between USB and BT for accessing the LEDs/OLEDs.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>