Input - wacom: use a specific name for the battery device
The current implementation uses "wacom_battery" as a generic name for
batteries. This prevents us to have two Wacom devices with a battery
attached as the power system will complain about the name which is already
registered.

Use a specific name for each type of battery attached.

A bug in upower prevents us to use spaces in the power name, so replace
each space by '_'.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>