HID: uclogic: apply quirk NO_EMPTY_INPUT
NO_EMPTY_INPUT is useful when MULTI_INPUT is set. It prevents to create
empty input nodes that user space does not know what to do with.

It does not seem to be required at the moment, this is just a preventive
patch. This check is only made during the plug of the device, so it does
not hurt to have it.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>