Input - rmi_smbus: fix after reset lost of the device
Once the device resets, the ps/2 driver takes over and prevents the SMBus
driver to talk to the device until it has discarded the touchpad.

Wait for 1500 ms for the ps/2 driver to end, and also discard the current
SMBus state.
We also need to querry the SMBus version of the touchpad to have it bound
on the SMBus.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>