Input - rmi4: embed the function modules in rmi_core
the function modules can not be auto-loaded by udev. So at boot, the
functions are not there and the touchpad does not answer.
Force the module to be embedded in rmi_core so that when the touchpad
is there, the functions are there too.

There is not much use of having the functions separate anyway

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>