i2c: i801: add support of Host Notify
The i801 chip can handle the Host Notify feature (at least from ICH 7 as
mentioned in
http://www.intel.com/content/dam/doc/datasheet/i-o-controller-hub-7-datasheet.pdf)

Implement .toggle_smbus_host_notify() and rely on .alert() to notify
the actual client that it has a notification.

With a T440s and a Synaptics touchpad that implements Host Notify, I can
not seem to access the payload data from the Host Notify packet, so this
is left aside right now.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>