Input - rmi4: enable suspend/resume for SMBus devices
In rmi_driver, we need to call rmi_driver_process_config_requests() to
enable the interrupts to be sent by the device.
That means that we can skip this part in .probe() given that it will be
called in .enable().

For the SMBus side, we need to clear the current state, be sure that the
device is bound and call the function $FD to enable the SMBus protocol.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>