Input: psmouse - allow synaptics_init to abort and not bind the PS/2 device
Most of the Synaptics devices are connected through PS/2 and a different
bus. Most of the time, we do not want the PS/2 connection to be used,
but if the ps/2 initialization goes too far, the other bus is disconnected
by the device.
Allow to abort when synaptics_init() fails so that we keep the other bus
alive.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>