Input: synaptics - handle rescan while using SMBus
On resume, the PS/2 synaptics driver gets called with a rescan.
This rescan switches back the touchpad in the PS/2 mode, and we
need to re-enable the SMBus mode. Rely on the .resume() callback
to notify the device that it needs to clear its state.

Doing a rescan of the PS/2 port causes a segfault, due to the fact the
i2c device gets unbound but never removed, meaning we try to call a
resume function from a non-existant driver. Easiest fix is unregistering
the device when we find the driver's been unbound, since this lets us
bring the device back up on smbus and avoid segfaulting.

Signed-off-by: Stephen Chandler Paul <cpaul@redhat.com>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>