Input: synaptics - create the InterTouch SMBus device from the PS/2 driver
The PS/2 driver knows when we should instantiate the SMBus device, so
use it to create the various devices.
We only enable the Intertouch device to be created for the laptops
registered with the top software button property or those we know
they are functional.

In the future, we might change the default to always rely on the
InterTouch bus. Currently, users can enable/disable the feature
with the psmouse parameter synaptics_intertouch.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>