Input: synaptics-rmi4 - Ignore interrupts on suspended sensors
Sometimes while trying to reenable sensors when resuming from standby,
we run into an issue where the interrupt handler catches an interrupt
from an RMI4 device despite the fact the device isn't actually ready to
communicate with yet. As such, this causes the interrupt handler to fail
and print some annoying messages to the kernel log.

Signed-off-by: Stephen Chandler Paul <cpaul@redhat.com>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>