Input: synaptics-rmi4 - Fix suspend/resume by not using pm_ops
In order to resume RMI4 devices properly, we first have to bring them
back up on the PS/2 host, only after doing this can we actually begin
any of the actual work done to resume the devices. Now, we instead have
the psmouse handle the resume work of the RMI4 device.

Signed-off-by: Stephen Chandler Paul <cpaul@redhat.com>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>