Input: synaptics-rmi4 - Don't try to resume an unbound device
Doing a rescan of the PS/2 port causes a segfault, due to the fact the
i2c device gets unbound but never removed, meaning we try to call a
resume function from a non-existant driver. Easiest fix is unregistering
the device when we find the driver's been unbound, since this lets us
bring the device back up on smbus and avoid segfaulting.

Signed-off-by: Stephen Chandler Paul <cpaul@redhat.com>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>