HID: wacom: mark generic hid indirect devices as indirect
Generic HID devices used to be direct. Now that we are injecting custom
firmwares, we have generic HID devices that are indirect too.
We rely on the HID_DG_TOUCHPAD application for the touch devices and
whether there is the mouse capability on the pen ones to detect such
indirect devices.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>