HID: wacom: sideload an external hid report descriptors
put to sleep to avoid warn slow path
 -> not anymore???

describe the fw file format

put description in Documentation too

add version numbering (might be just a string from the tool
that generated the firmware and showed as an info).

add checksum control

HID: wacom: be less restrictive regarding the tags of the firmwares

Do not abort if an unknown tag is seen.
Allow for further tags to be added.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>

HID: wacom: retrieve the version in the firmware

The version is just a string that needs to be shown in the dmesg for
debugging problems.
This makes the version tag mandatory in the firmware file.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>