HID: wacom: postpone the loading of the firmware when it's not found
If the firmware is not found in .probe(), it's either that the firmware
is not available at all, or that we are probing while in the initrd.
In the latter case, the firmware might become available later, when the
filesystem gets mounted.

Unfortunately, latest versions of udev removed the ability to provide
a firmware when it receives a kevent, so we need to create a blank input
device which will be opened once the filesystem is ready, and which will
in turn check for the firmwares and load the correct input nodes.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>