HID: wacom: add support of generic report descriptors for Wacoms
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>

HID: wacom: add Z handling for Bamboos pen

Bamboos are actually following the HID spec. The reported distance from
the surface is given through the Z axis, which is facing downward.

Add a note on ALTITUDE that we are actually misusing a HID usage and why
we need to invert the Z axis.

This patch will be squashed with "HID: wacom: add support of generic
report descriptors for Wacoms"

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>

HID: wacom: add support of pads from the hid collection

When a collection is declared to be the one of the pad, we should map
the fields accordingly.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>

HID: wacom: add support of touches from the hid collection

Like the pen and pad, we can also inject in the report descriptor the
semantic of the touches.

Few things here to note that may impact generic touch devices:
- the devices still declares as direct (this will be fixed in a later
  commit)
- wacom_wac->features.touch_max is overwritten by the range of Contact
  Count
- if the contact ID is outside the logical range, we ignore the contact

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>

HID: wacom: fix touch_max

for finger usage_mapping

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>

HID: wacom: break out custom usage handling

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>

HID: wacom: introduce WACOM_REPLACE_BITS to save repetitive writing

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>

HID: wacom: remove duplicated code in *usage_mapping

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>