HID: wacom: mark generic hid indirect devices as indirect
Current Wacom generic HID devices are direct. However, for being able to
support generic indirect devices, we need to handle the indirect case
properly.
For the touch devices, the most sensible information we have is that it
should declare the HID_DG_TOUCHPAD application. For the pen devices,
the indirect devices support mice while direct don't. This mouse detection
is actually a leap forward to the next wacom update, but it feels OK
having it here as it won't break any existing devices.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>