HID: wacom: create input nodes for handled devices only
Commit 042628abd59 ("HID: wacom: Discover device_type from HID descriptor
for all devices") marked the devices presenting WACOM_VENDORDEFINED_PEN
as an application usage as a pen device.

However, in the generic HID case, this collection does not present any
standard HID usages, and gets unpopulated.

Before actually asking the driver to present an empty input node, make
sure the pen and touch input nodes are populated with absolute axis, which
is a good heuristic that the tool has a chance to be handled.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>