HID: wacom: allow wireless devices to use the firmwares
There is a slight difference with regular firmware matching:
the original report descriptor is not valid for use as base, so we need
to specifically request a tag in the list of fixed report descriptors.
'T' and 'O' are mutually exclusive.

In wireless_work, we now need to stop and restart the hid device in order
to parse the report descriptor. Given that we might end up in a situation
where the device is stopped but not restarted, we now check whether or not
the device has been started/stopped to prevent any double start or double
stop.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>