Input: psmouse - allow synaptics_init to abort and not bind the PS/2 device
Most of the Synaptics devices are connected through PS/2 and a different
bus. Most of the time, we do not want the PS/2 connection to be used,
but if the ps/2 initialization goes too far, the other bus is disconnected
by the device.
Allow to abort when synaptics_init() fails so that we keep the other bus
alive.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>

Input: psmouse - fix allow synaptics to unbind

Input: synaptics - abort the initialization when there is an other bus available

When the synaptics PS/2 driver detects that the device can be connected
through a different bus, abort the initialization and wait for the other
bus to handle the device.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>

Input: synaptics - get notified when a SMBus adapter is available

Will be used to instanciate a SMBus device directly from PS/2.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>

Input: synaptics - create the InterTouch SMBus device from the PS/2 driver

The PS/2 driver knows when we should instantiate the SMBus device, so
use it to create the various devices.
We only enable the Intertouch device to be created for the laptops
registered with the top software button property or those we know
they are functional.

In the future, we might change the default to always rely on the
InterTouch bus. Currently, users can enable/disable the feature
with the psmouse parameter synaptics_intertouch.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Andrew Duggan <aduggan@syanptics.com>