Input: synaptics - export synaptics_send_ps2_cmd()
For RMI4, we need to have a separate PS/2 driver. This driver needs
to enumerate the capabilities over PS/2 as well as there are some
that are only available under PS/2, and not SMBus (trackstick buttons
on the T450s, I am looking at you).

Instead of duplicating the code, export the internal synaptics send
command and strip the psmouse specific bits.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>