HID: wacom: use the ledclass instead of custom made sysfs files
The now obsolete sysfs files for LEDs and EKRemote are kept for backward
compatibility.
Both the EKR (read-only) and the regular Cintiqs and Intuos are now
sharing the same led API.
The switch of the LED is now handled by the kernel, which allows a common
behavior between the EKR and the other Wacoms. The other benefit from this
is that user-space don't need to have root privileges to read the LED
state.

Warning when unpairing an EKR, though need further work

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>