HID: wacom: EKR: have one power_supply per remote
Previously, all the remotes attached to the same receiver would share the
same power_supply. That's not good as the remotes will constantly change
the battery information according to their own state.

To have something generic enough, we introduce struct wacom_battery
which regroups all the informations we need for a battery.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>