i2c: i2c-smbus: fix return value of i2c_handle_smbus_host_notify()
schedule_work() returns a boolean (true on success, false if the work
was already queued).
This doesn't match the "negative number on error" return model that the
function exports.

Given that schedule_work() will always return true (we have an internal
.pending check protected by a spinlock), just return false and ignore
the return value of schedule_work().

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>

---

no changes in v3

new in v2