HID: rmi: fix rmi4 irq reporting
Using handle_nested_irq() prevents a warning to be sent but doesn't
make use of the threading capability of the IRQ. So we need to use
generic_handle_irq() with an IRQ lock. Furthermore, the "current"
implementation is racy as the transport driver writes fields in various
places of rmi_driver. Better have a proper API for it which provides
a lock.

One remaining issue is that if 2 IRQs happen while only one is getting
processed, we might lose one IRQ and probably valuable information.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>