tty/vt/keyboard: reset the LEDs state at each console change
When switching between consoles, the LEDs state is correctly
assigned, as long as the kernel manages the console.
When switching back from Gnome, the VT is not aware of the
current state of the LEDs. So if Gnome changes them, the
kernel still bellieves they are off, and it won't turn them
off when switching to a new TTY.

To keep the LEDs status in sync with the actual modifiers,
simply force a reset of the LEDs when we detect a vt change.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>