Input: elantech - add support for SMBus devices
Most of the Elantech devices are connected through PS/2 and a different
bus (SMBus).

We only enable the InterTouch device to be created for the laptops
we know are functional. In the future, we might change the default
to always rely on the SMBus bus.
Currently, users can enable/disable the feature with the psmouse
parameter elantech_smbus.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>