platform/x86: surface3_power: MSHW0011 rev-eng implementation
The MSHW0011 device is a chip that replaces the battery firmware
by using ACPI operation regions on the Surface 3.
It is unclear whether or not the chip will be reused somewhere else
(under Windows, the chip is called "Surface Platform Power Driver"
and the driver is provided by Microsoft).

The values have been obtained by reverse engineering, and are subject to
errors. Looks like it works on overall pretty well.

I couldn't manage to get the IRQ correctly triggered, so I am using a
good old polling thread to check for changes. This is something
to be fixed in a later version.

Link: https://bugzilla.kernel.org/show_bug.cgi?id=106231

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Stephen Just <stephenjust@gmail.com>

---

changes in v2:
- moved to drivers/acpi/ instead of power
- use uuid_le
- fix uper/lower case
- print_hex_dump() used in mshw0011_dump_registers() instead of custom dump
- removed "MSHW0011:00" in mshw0011_id

changes in v3:
- remove le16_to_cpu() after i2c_smbus_read_word_data() as it already
  handles it properly
- use i2c_acpi_new_device() to remove a bunch of duplicated code from I2C
  core.
- move the driver in platform/x86
- add depends ACPI && I2C in Kconfig
- remove the dump registers facility, as it's debugging only
- use of BIT() in MSHW0011 defs
- use of guid_t
- remove useless ret = 0
- use probe_new()
- remove empty .id_table
- removed mshw0011_i2c_read_block() and use directly
  i2c_smbus_read_i2c_block_data()
- use snprintf() instead of custom memcpy()
- renamed the 'version' as 'mask'
- use SPDX license identifier