HID: force setting drvdata to NULL when removing the driver
Or when the probe failed.

This is a common pattern in the HID drivers to reset the drvdata. Some
do it properly, some do it only in case of failure.
Anyway, it's never a good thing to have breadcrumbs, so force a clean
state when removing or when probe is failing.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>